Rename subpackage trk -> nav (#180)
