+++ src/openlifu/nav/meshroom_pipelines/default_pipeline.mg
