+++ src/openlifu/nav/meshroom_pipelines/downsample_4x_pipeline.mg
