+++ src/openlifu/nav/meshroom_pipelines/downsample_8x_pipeline.mg
