+++ src/openlifu/nav/meshroom_pipelines/draft_pipeline.mg
